Password Settings
On an MVC Application  on the IdentityConfig.cs class under the App_Start folder, where the code is based or should be based on your Application.
You can manipulate the Password or Set it the way that you want to, but make sure that the Password that you are setting is a Strong Password. The given code has the maximum length of 6, also requires a special character, uppercase, lower case and also a digit and if your password does not include one of the requirements, when the Application is executed, there will be a validation message that  is going to appear on your registration or log in page reminding/telling you on what is required.